requirements.txt
i2cylib>=1.11.1
matplotlib>=3.5.3
paho-mqtt
